Fix post-login broker bootstrap handling
