+++ apps/server/src/domain_events/queue_worker.rs
@@ -10,7 +10,8 @@ use std::time::Duration;
 
 use tokio::sync::mpsc;
 use wealthfolio_connect::{
-    ensure_valid_access_token, BrokerSyncServiceTrait, TokenLifecycleConfig, TokenLifecycleState,
+    acquire_broker_sync_guard, ensure_valid_access_token, BrokerSyncServiceTrait,
+    TokenLifecycleConfig, TokenLifecycleState,
 };
 use wealthfolio_core::{
     assets::AssetServiceTrait,
@@ -61,25 +62,6 @@ pub struct QueueWorkerDeps {
     /// Categorization rules service — auto-runs rules against newly-changed activities.
     pub categorization_rules_service:
         Arc<wealthfolio_spending::categorization_rules::CategorizationRulesService>,
-}
-
-struct BrokerSyncRunGuard {
-    running: Arc<AtomicBool>,
-}
-
-impl Drop for BrokerSyncRunGuard {
-    fn drop(&mut self) {
-        self.running.store(false, Ordering::Release);
-    }
-}
-
-fn try_acquire_broker_sync_guard(running: &Arc<AtomicBool>) -> Option<BrokerSyncRunGuard> {
-    running
-        .compare_exchange(false, true, Ordering::AcqRel, Ordering::Acquire)
-        .ok()
-        .map(|_| BrokerSyncRunGuard {
-            running: Arc::clone(running),
-        })
 }
 
 /// Runs the event queue worker.
@@ -269,7 +251,7 @@ async fn process_event_batch(events: &[DomainEvent], deps: Arc<QueueWorkerDeps>)
         let broker_sync_running = deps.broker_sync_running.clone();
 
         tokio::spawn(async move {
-            let Some(_guard) = try_acquire_broker_sync_guard(&broker_sync_running) else {
+            let Some(_guard) = acquire_broker_sync_guard(&broker_sync_running) else {
                 tracing::info!(
                     "Broker sync skipped after tracking mode change: sync already running"
                 );
